+++ mariadb/CHANGELOG.md
@@ -7,6 +7,7 @@
 * Fix error in changing mysql-server rc script
 * Add GALERAHOST parameter to set an access rule for a loadbalancer host
 * Rename GALERAHOST to LOADBALANCER, galera user to haproxy, to avoid confusion with galera package
+* Add relay-log to avoid replication problems on hostname change
 
 ---