+++ mariadb/CHANGELOG.md
@@ -1,3 +1,9 @@
+3.9
+
+* Version bump for new base image
+
+---
+
 3.8
 
 * Version bump for new base image